+++ age/CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to Rust's notion of
 to 1.0.0 are beta releases.
 
 ## [Unreleased]
+### Added
+- `age::armor::ArmoredReader`, which can be wrapped around an input to handle
+  a potentially-armored age file.
+- `age::armor::ArmoredWriter`, which can be wrapped around an output to
+  optionally apply the armored age format.
+
+### Changed
+- `age::Encryptor::wrap_output` now only generates the non-malleable binary age
+  format. Use `encryptor.wrap_output(age::armor::ArmoredWriter(output, format))`
+  to optionally generate armored age files.
+- `age::Decryptor` now only decrypts the non-malleable binary age format. Use
+  `age::Decryptor::new(age::armor::ArmoredReader::new(input))` to handle age
+  files that are potentially armored.
+- `age::Format` has been moved to `age::armor::Format`.
 
 ## [0.4.0] - 2020-03-25
 ### Added